hospital-management/apps/api/pyproject.toml
[project]
name = "hospital-api"
version = "0.1.0"
description = "Hospital management backend API"
requires-python = ">=3.12"
dependencies = [
  "alembic==1.16.5",
  "celery==5.5.3",
  "fastapi==0.116.1",
  "httpx==0.28.1",
  "python-jose[cryptography]==3.5.0",
  "psycopg[binary]==3.2.9",
  "passlib[bcrypt]==1.7.4",
  "pydantic-settings==2.10.1",
  "redis==6.4.0",
  "sqlalchemy==2.0.43",
  "uvicorn[standard]==0.35.0",
]

[dependency-groups]
dev = [
  "mypy==1.17.1",
  "pytest==8.4.1",
  "ruff==0.12.11",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = false
ignore_missing_imports = true
pretty = true
warn_redundant_casts = true
warn_unused_ignores = true
